Rework diplacement filter to sample-based approach (#3311)
Refactoring:
* Get rid of unnecessary DataDependantSetup
* Introduce SetupImpl
* Change pass-by-pointer to pass-by-reference
* Rework GPU Op to process flattened blocks instead
  of whole images accessed via offset to one global
  batch pointer.
* Instead of accessing underlying contiguous
  TL buffer we access individual samples in GPU Op.
* Rework CPU Op to use HostWorkspace
* Keep the optimized implementation for aligned 
  samples.

Fixes:
* Fix masking in CPU Op - access it as int instead
  of bool to be consistent with the default and GPU Op.
* Fix masking in CPU Op - copy does not try to access
  stream in CPU workspace which causes error.

Tests:
* Add paths in water test to cover more code paths
   * prime-sized image to fall into the non-optimized
     kernel
   * mask support
   * both input types: uint8 and float32.

Other adjustments:
* Adjust block setup for 1 dim to {256, 1, 1} block.

+++ dali/operators/generic/constant.cu
@@ -65,7 +65,7 @@ void FillTensorList(
                                                      opaque(ConvertSat<Dst>(src[0])));
   } else {
     SmallVector<Dst, 64> tmp;
-    assert(src.size() == sample_size);
+    assert(static_cast<int>(src.size()) == sample_size);
     tmp.resize(src.size());
     for (size_t i = 0; i < tmp.size(); i++)
       tmp[i] = ConvertSat<Dst>(src[i]);
